update envVar from workflow
